Create grid for Etch-A-Sketch game
Removed initial JS code as too many divs needed to be added to the DOM
to create the grid in the first place. Webpage would also crash when
trying to load. New code utilizes a function with a for loop so the
page doesn't crash on load, creating cleaner code and a better way for
the player to create different size grids at scale in the future .
Reworked CSS to style the grid and get it to actually show on the
webpage unlike the previous method as well as show in the proper
middle position. Renamed the website title in HTML to better reflect
what's on the webpage.