Modify Etch-A-Sketch grid styling
Removed the numbered text on each cell of the grid. This makes the grid
less cluttered once the hover mouse function is added. Resized the
padding of each cell to get as close to it's original size as possible.
This was needed due to the removal of the numbers making the cells
smaller