Add a reset button to etch-a-sketch
Added a button called "Reset" with the intent of removing the
existing grid and creating a new grid with the parameters that the
player would input in a prompt when the button was clicked. The
button currently does not work properly however. Repositioned the
grid in order to place the button at the top of the page.