Create a new grid with the reset button
The reset button now correctly creates a new blank grid of the size
that the user entered and keeps it in grid format instead of changing
the grid to a column. The orientation of the grid changes however.
Basically the new grid gets pushed to the right for any grid larger
than 16 x 16. The larger the new grid, the further the grid is pushed
to the right on the page. Working on a fix for this.