Fix grid errors and modify grid UI
Changed the grid layout as the previous layout would not keep larger
grids being made in the viewport. Scroll bars would show up. Also
lowered the limit of new grids to 80 x 80 for the same reason. Reduced
the size of the reset button for the same reason. These changes keep
everything within the viewport regardless of the size of the grid.